primary-display/pyproject.toml
[project]
name = "mgb-primary-display"
version = "0.1.0"
description = "MGB Dash 2026 — Primary Display (Pi 4B + Waveshare 3.4\" Round DSI LCD)"
requires-python = ">=3.9"
dependencies = [
    "pycairo>=1.24",
    "pygame>=2.5",
    "python-can>=4.0",
    "numpy>=1.24",
]

[project.scripts]
mgb-primary-display = "main:main"
